+++ packages/happn-cluster/test/lib/hooks.js
@@ -1,26 +1,27 @@
 process.env.NODE_TLS_REJECT_UNAUTHORIZED = 0;
 
-var path = require('path');
-var ChildProcess = require('child_process');
-var clone = require('clone');
-var HappnCluster = require('../../');
-var testUtils = require('./test-utils');
+let path = require('path');
+let ChildProcess = require('child_process');
+let clone = require('clone');
+let HappnCluster = require('../../');
+let testUtils = require('./test-utils');
 const wait = require('await-delay');
 const test = require('happn-commons-test').create();
 
 module.exports.startCluster = function (clusterOpts) {
-  var testSequence = clusterOpts.testSequence || 1;
-  var clusterSize = clusterOpts.size || 5;
-  var happnSecure = typeof clusterOpts.happnSecure === 'boolean' ? clusterOpts.happnSecure : false;
-  var proxySecure = typeof clusterOpts.proxySecure === 'boolean' ? clusterOpts.proxySecure : false;
-  var services = clusterOpts.services || {};
-  var clusterConfig = clusterOpts.clusterConfig;
+  let testSequence = clusterOpts.testSequence || 1;
+  let clusterSize = clusterOpts.size || 5;
+  let happnSecure = typeof clusterOpts.happnSecure === 'boolean' ? clusterOpts.happnSecure : false;
+  let proxySecure = typeof clusterOpts.proxySecure === 'boolean' ? clusterOpts.proxySecure : false;
+  let services = clusterOpts.services || {};
+  let clusterConfig = clusterOpts.clusterConfig;
+  let additionalDatastore = clusterOpts.datastore;
   before('clear collection (before)', function (done) {
     testUtils.clearMongoCollection(done);
   });
 
   before('start cluster', async function () {
-    var self = this;
+    let self = this;
 
     if (!clusterConfig)
       self.__configs = await testUtils.createMemberConfigs(
@@ -39,10 +40,15 @@ module.exports.startCluster = function (clusterOpts) {
         services,
         clusterConfig
       );
+    if (additionalDatastore) {
+      self.__configs.forEach((config) => {
+        config.services.data.config.datastores.push(additionalDatastore);
+      });
+    }
     let servers = [];
     servers.push(HappnCluster.create(clone(self.__configs[0])));
     await test.delay(2000);
-    // start first peer immediately and other a momentf
+    // start first peer immediately and other a moment
     // later so they don't all fight over creating the
     // admin user in the shared database
     for (let [sequence, config] of self.__configs.entries()) {
@@ -79,16 +85,16 @@ module.exports.startMultiProcessCluster = function (clusterOpts) {
     testUtils.clearMongoCollection(done);
   });
 
-  var testSequence = clusterOpts.testSequence || 1;
-  var clusterSize = clusterOpts.size || 5;
-  var happnSecure = typeof clusterOpts.happnSecure === 'boolean' ? clusterOpts.happnSecure : false;
-  var proxySecure = typeof clusterOpts.proxySecure === 'boolean' ? clusterOpts.proxySecure : false;
-  var services = clusterOpts.services || {};
+  let testSequence = clusterOpts.testSequence || 1;
+  let clusterSize = clusterOpts.size || 5;
+  let happnSecure = typeof clusterOpts.happnSecure === 'boolean' ? clusterOpts.happnSecure : false;
+  let proxySecure = typeof clusterOpts.proxySecure === 'boolean' ? clusterOpts.proxySecure : false;
+  let services = clusterOpts.services || {};
 
-  var peerPath = __dirname + path.sep + 'peer.js';
+  let peerPath = __dirname + path.sep + 'peer.js';
 
   before('start cluster', async function () {
-    var self = this;
+    let self = this;
 
     self.__configs = await testUtils.createMemberConfigs(
       testSequence,
@@ -100,10 +106,10 @@ module.exports.startMultiProcessCluster = function (clusterOpts) {
 
     let processes = await Promise.all(
       self.__configs.map(function (config) {
-        var configJson = [JSON.stringify(config)];
+        let configJson = [JSON.stringify(config)];
 
         return new Promise(function (resolve) {
-          var peerProcess = ChildProcess.fork(peerPath, configJson);
+          let peerProcess = ChildProcess.fork(peerPath, configJson);
           peerProcess.on('message', function (message) {
             if (message === 'ready') return resolve(peerProcess);
           });
@@ -127,5 +133,13 @@ module.exports.stopMultiProcessCluster = function () {
 
   after('multi clear collection (after)', function (done) {
     testUtils.clearMongoCollection(done);
+  });
+};
+module.exports.clearNamedCollection = function (collectionName) {
+  before('clear collection (before)', function (done) {
+    testUtils.clearMongoCollection(done, collectionName);
+  });
+  afterEach('clear collection (after)', function (done) {
+    testUtils.clearMongoCollection(done, collectionName);
   });
 };