+++ packages/happner-2/test/scratch/timer/index.js
@@ -1,8 +1,0 @@
-module.exports = class Timer {
-  constructor() {}
-  async start() {}
-  async getSystemTime() {
-    let time = Date.now();
-    return time;
-  }
-};