+++ packages/happner-cluster/test/_lib/integration-37-broker-component/index.js
@@ -1,0 +1,19 @@
+module.exports = Component;
+
+function Component() {}
+
+Component.prototype.start = function ($happn, callback) {
+  callback();
+};
+
+Component.prototype.stop = function ($happn, callback) {
+  callback();
+};
+
+Component.prototype.block = function ($happn, callback) {
+  setTimeout(() => {
+    const target = Date.now() + 10000;
+    while (Date.now() <= target) {}
+  }, 100);
+  callback(null, $happn.info.mesh.name + ':brokerComponent:block');
+};