+++ packages/happn-3/RELEASES.md
@@ -966,7 +966,19 @@
   - TEN-103: set default timeout on component loading wait warning to 30secs
   - TEN-104: errors are only logged on system failures
 
-13.0.7 2022-04-18
+13.0.7 2022-05-04
+-----------------
+  - TEN-112: group with permissions that have dots in them causes compaction failure
+
+13.0.8 2022-05-06
+-----------------
+  - TEN-114: enforce admin group save on startup
+
+13.0.9 2022-05-09
+-----------------
+  - TEN-114: happner-2: enforce MESH_GST, MESH_ADMIN upsert on startup, ensures db dir
+
+13.0.10 2022-05-17
 -----------------
   - SMC-4817: mongo-based membership strategy