+++ packages/happner-cluster/RELEASES.md
@@ -359,6 +359,23 @@
 -----------------
   - SMC-4817: mongo-based membership strategy
 
+12.1.2 2022-05-04
+-----------------
+  - TEN-112: happn-3: group with permissions that have dots in them causes compaction failure
+
+12.1.3 2022-05-06
+-----------------
+  - TEN-114: happn-3: enforce admin group save on startup
+
+12.1.4 2022-05-09
+-----------------
+  - TEN-114: happner-2: enforce MESH_GST, MESH_ADMIN upsert on startup
+
+12.1.5 2022-05-17
+-----------------
+  - SMC-4817: mongo-based membership strategy
+
+